requirements.txt
Appium-Python-Client==0.28
atomicwrites==1.1.5
attrs==18.1.0
more-itertools==4.2.0
pluggy==0.7.1
py==1.5.4
pytest==3.7.0
selenium==3.13.0
six==1.11.0